Updated the RSS feed to typescript, and to use the new content loading
